+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * (theimo1221) Add Humidity Sensor Persisting
 * (theimo1221) Clear Persisting from old mixed persisting commands
 * (theimo1221) Make battery readonly to outside
+* (theimo1221) Persist Battery Device Data
 
 ## 1.1.48 (2022-10-15)